Switched floats to ints in gradient shaders for better performance (#40189)
Switched floats to ints in gradient shaders for better performance

+++ impeller/entity/shaders/linear_gradient_ssbo_fill.frag
@@ -21,7 +21,7 @@ uniform FragInfo {
   vec2 end_point;
   float alpha;
   float tile_mode;
-  float colors_length;
+  int colors_length;
 }
 frag_info;
